+++ src/main/java/com/mercadodabola/mercadotransferencia/domain/util/CalculaIdade.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class CalculaIdade {
 	
-	public Long getIdade(LocalDate dataNascimento) {
+	public Integer getIdade(LocalDate dataNascimento) {
 		Date dateNascimento = Date.from(dataNascimento.atStartOfDay().atZone(ZoneId.systemDefault()).toInstant());
 		Date dataAtual = new Date();
 
@@ -18,12 +18,12 @@ public class CalculaIdade {
 
 		Long diferencaDatas = dateAtualMs - dateNascimentoMs;
 		
-		return diferencaDatas / 1000 / 60 / 60 / 24 / 365;
+		return (int) (diferencaDatas / 1000 / 60 / 60 / 24 / 365);
 		
 	}
 	
 	public Boolean isIdadeValida(LocalDate dataNascimento) {
-		Long idade = getIdade(dataNascimento);
+		Integer idade = getIdade(dataNascimento);
 		if (idade < 18 || idade >= 30) {
 			return false;
 		}
